refactor(split-legacy): Moved filter handlers to core module and reenable deprecated getSegmentContacts method
