+++ apps/backend/src/api/transformers/CalloutTransformer.ts
@@ -24,12 +24,13 @@ import {
 import { BaseTransformer } from "@api/transformers/BaseTransformer";
 import CalloutVariantTransformer from "@api/transformers/CalloutVariantTransformer";
 import { groupBy } from "@api/utils";
-import { mergeRules, statusFilterHandler } from "@beabee/core/utils/rules";
+import { mergeRules } from "@beabee/core/utils/rules";
 
 import { Callout, CalloutResponse, CalloutVariant } from "@beabee/core/models";
 
 import { AuthInfo, FilterHandlers } from "@beabee/core/type";
 
+import { calloutFilterHandlers } from "@beabee/core/filter-handlers";
 class CalloutTransformer extends BaseTransformer<
   Callout,
   GetCalloutDto,
@@ -39,13 +40,8 @@ class CalloutTransformer extends BaseTransformer<
   protected model = Callout;
   protected modelIdField = "id";
   protected filters = calloutFilters;
-  protected filterHandlers: FilterHandlers<CalloutFilterName> = {
-    status: statusFilterHandler,
-    title: (qb, args) => {
-      // Filter by variant title
-      qb.where(args.convertToWhereClause("cvd.title"));
-    }
-  };
+  protected filterHandlers: FilterHandlers<CalloutFilterName> =
+    calloutFilterHandlers;
 
   protected async transformFilters(
     query: GetCalloutOptsDto & PaginatedQuery,
