+++ apps/backend/src/api/transformers/NoticeTransformer.ts
@@ -7,7 +7,8 @@ import { TransformPlainToInstance } from "class-transformer";
 
 import { BaseTransformer } from "@api/transformers/BaseTransformer";
 import { GetNoticeDto, ListNoticesDto } from "@api/dto/NoticeDto";
-import { mergeRules, statusFilterHandler } from "@beabee/core/utils/rules";
+import { mergeRules } from "@beabee/core/utils/rules";
+import { statusFilterHandler } from "@beabee/core/filter-handlers";
 
 import { Notice } from "@beabee/core/models";
