+++ packages/core/package.json
@@ -24,12 +24,15 @@
     "./config": "./dist/config/config.js",
     "./database": "./dist/database.js",
     "./errors": "./dist/errors/index.js",
+    "./filter-handlers": "./dist/filter-handlers/index.js",
+    "./filter-handlers/*": "./dist/filter-handlers/*.js",
     "./locale": "./dist/locales/current.js",
     "./logging": "./dist/logging.js",
     "./models": "./dist/models/index.js",
     "./type": "./dist/type/index.js",
     "./server": "./dist/server.js",
     "./lib/*": "./dist/lib/*.js",
+    "./services": "./dist/services/index.js",
     "./services/*": "./dist/services/*.js",
     "./utils/*": "./dist/utils/*.js"
   },
