+++ packages/core/src/filter-handlers/README.md
@@ -1,0 +1,39 @@
+# Filter Handlers
+
+The filter handlers are responsible for managing and applying filters to data transformations within the beabee application. They provide a standardized way to filter and transform data models before they are sent to the client.
+
+## Purpose
+
+- Provide a centralized location for all filter logic
+- Ensure consistent data transformation across the application
+- Allow for flexible and reusable filtering mechanisms
+- Support type-safe filtering operations
+
+## Usage
+
+Filter handlers are typically used in conjunction with transformers to modify or filter data based on specific conditions. They can be applied to any data model and are especially useful when:
+
+- Converting database models to DTOs
+- Filtering sensitive information based on user roles
+- Implementing conditional data transformation
+
+## Example
+
+```typescript
+class ExampleTransformer extends BaseTransformer<Model, Dto> {
+  protected filterHandlers = {
+    // Define filters that should be applied during transformation
+    removePrivateData: (data: Model) => {
+      // Filter implementation
+    }
+  };
+}
+```
+
+## Best Practices
+
+1. Keep filters pure and focused on a single responsibility
+2. Document the purpose and expected behavior of each filter
+3. Consider performance implications when implementing filters
+4. Use TypeScript for type safety
+5. Test filters thoroughly with different scenarios
