+++ packages/core/src/utils/rules.ts
@@ -4,7 +4,6 @@ import {
   validateRuleGroup,
   ValidatedRule,
   ValidatedRuleGroup,
-  ItemStatus,
   RuleOperator,
   FilterType,
   operatorsByType,
@@ -26,10 +25,12 @@ import {
 } from "typeorm";
 import { QueryDeepPartialEntity } from "typeorm/query-builder/QueryPartialEntity";
 
+import { simpleFilterHandler } from "#filter-handlers";
+
 import { createQueryBuilder } from "#database";
 import { Contact } from "#models";
 
-import {
+import type {
   FilterHandler,
   FilterHandlers,
   RichRuleValue,
@@ -109,49 +110,6 @@ const operatorsWhereByType: Record<
     ...equalityOperatorsWhere,
     ...nullableOperatorsWhere
   })
-};
-
-// Generic field handlers
-
-const simpleFilterHandler: FilterHandler = (qb, args) => {
-  qb.where(args.convertToWhereClause(`${args.fieldPrefix}${args.field}`));
-};
-
-/**
- * Status is a virtual field that maps to starts and expires, this function
- * applies the correct filter for the status field value
- *
- * @param qb The query builder
- * @param args The rule arguments
- */
-export const statusFilterHandler: FilterHandler = (qb, args) => {
-  // TODO: handle other operators
-  if (args.operator !== "equal") {
-    throw new BadRequestError("Status field only supports equal operator");
-  }
-
-  switch (args.value[0]) {
-    case ItemStatus.Draft:
-      qb.where(`${args.fieldPrefix}starts IS NULL`);
-      break;
-    case ItemStatus.Scheduled:
-      qb.where(`${args.fieldPrefix}starts > :now`);
-      break;
-    case ItemStatus.Open:
-      qb.where(`${args.fieldPrefix}starts < :now`).andWhere(
-        new Brackets((qb) => {
-          qb.where(`${args.fieldPrefix}expires IS NULL`).orWhere(
-            `${args.fieldPrefix}expires > :now`
-          );
-        })
-      );
-      break;
-    case ItemStatus.Ended:
-      qb.where(`${args.fieldPrefix}starts < :now`).andWhere(
-        `${args.fieldPrefix}expires < :now`
-      );
-      break;
-  }
 };
 
 // Rule parsing
